separates label description
