+++ composeml/label_times/object.py
@@ -4,7 +4,8 @@ import os
 import pandas as pd
 
 from .data_frame import DataFrame
-from .label_plots import LabelPlots
+from .description import describe_label_times
+from .plots import LabelPlots
 
 
 class LabelTimes(DataFrame):
@@ -85,42 +86,7 @@ class LabelTimes(DataFrame):
         return value
 
     def describe(self):
-        """Prints out label info with transform settings that reproduce labels."""
-        if self.label_name is not None and self.is_discrete:
-            print('Label Distribution\n' + '-' * 18, end='\n')
-            distribution = self[self.label_name].value_counts()
-            distribution.index = distribution.index.astype('str')
-            distribution.sort_index(inplace=True)
-            distribution['Total:'] = distribution.sum()
-            print(distribution.to_string(), end='\n\n\n')
-
-        settings = self.settings.copy()
-        settings.update(settings.pop('search_settings'))
-        settings = pd.Series(settings)
-        transforms = settings.pop('transforms')
-
-        print('Settings\n' + '-' * 8, end='\n')
-
-        if settings.isnull().all():
-            print('No settings', end='\n\n\n')
-        else:
-            settings.sort_index(inplace=True)
-            print(settings.to_string(), end='\n\n\n')
-
-        print('Transforms\n' + '-' * 10, end='\n')
-
-        for step, transform in enumerate(transforms):
-            transform = pd.Series(transform)
-            transform.sort_index(inplace=True)
-            name = transform.pop('transform')
-            transform = transform.add_prefix('  - ')
-            transform = transform.add_suffix(':')
-            transform = transform.to_string()
-            header = '{}. {}\n'.format(step + 1, name)
-            print(header + transform, end='\n\n')
-
-        if len(transforms) == 0:
-            print('No transforms applied', end='\n\n')
+        describe_label_times(self)
 
     def copy(self, **kwargs):
         """
